sw_apps:zynqmp_pmufw:Fixed 10.4 MisraC violation
MisraC 10.4 violation states: Both operands of an operator should have
                              the same essential type while performing
                              arithmetic operations

This patch fixes the rule violation by proper typecasting to the operands


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/xpfw_scheduler.c
@@ -122,7 +122,7 @@ void XPfw_SchedulerTickHandler(XPfw_Scheduler_t *SchedPtr)
 	SchedPtr->Tick++;
 	for (Idx = 0U; Idx < XPFW_SCHED_MAX_TASK; Idx++) {
 		/* Check if it this task can be triggered */
-		if (TRUE == is_task_active(SchedPtr, Idx)) {
+		if ((u32)TRUE == is_task_active(SchedPtr, Idx)) {
 			/* Mark the Task as TRIGGERED */
 			SchedPtr->TaskList[Idx].Status = XPFW_TASK_STATUS_TRIGGERED;
 		}
@@ -144,7 +144,7 @@ void XPfw_SchedulerProcess(XPfw_Scheduler_t *SchedPtr)
 			SchedPtr->TaskList[Idx].Status = XPFW_TASK_STATUS_DISABLED;
 			CallCount++;
 	                /* Remove the Non-Periodic Task */
-		        if (TRUE == is_task_non_periodic(SchedPtr, Idx)) {
+		        if ((u32)TRUE == is_task_non_periodic(SchedPtr, Idx)) {
 			        SchedPtr->TaskList[Idx].Callback = NULL;
 			}
 		}
